Major enhancements

+++ core/classes/bullets/bullet.py
@@ -3,7 +3,7 @@ from variables import entities
 
 
 class Bullet(Objects):
-    def __init__(self, x: int, y: int, w: int, h: int, image: str, speed:int, damage:int, direction: int=0):
+    def __init__(self, x: int, y: int, w: int, h: int, image: str, speed:int, damage:int, direction: float=0):
         super().__init__(x, y, w, h, image)
         self.direction = direction
         self.speed = speed
